+++ pi/nn/__init__.py
@@ -1,1 +1,5 @@
 from .modules import *
+from .parameter import (
+    UninitializedParameter as Parameter,
+    UninitializedBuffer as Buffer,
+)
