+++ requirements.txt
@@ -1,3 +1,2 @@
 PyYAML
 pyccolo
-plum-dispatch
