+++ tests/torch_mlir/xfail.py
@@ -4,20 +4,18 @@ PI_XFAIL_SET = {
     "FlattenDynamicModule_basic",
     "ElementwiseExpm1Module_basic",
     "FlattenRank0Module_basic",
+
     # unknown
     "UniformModule_basic",
     "BernoulliFloatModule_basic",
+
     # torchvision
     "IouOfModule_basic",
     "ResNet18Module",
     "ResNet18Module_basic",
     "ResNet18StaticModule_basic",
     "MobilenetV3Module_basic",
-    # deprecated _convolution signature
-    "_ConvolutionDeprecated2DAllFalseModule_basic",
-    "_ConvolutionDeprecated2DBenchmarkModule_basic",
-    "_ConvolutionDeprecated2DDeterministicModule_basic",
-    "_ConvolutionDeprecated2DCudnnModule_basic",
+
     # tuple returns
     "BernoulliTensorModule_basic",
     "AtenEmbeddingBagSumExample_basic",
@@ -41,7 +39,8 @@ PI_XFAIL_SET = {
     "DropoutTrainModule_basic",
     "ReturnThreeTensorFloat32_basic",
     "ReturnTwoTensorF32I64_basic",
-    # python type return
+
+    # python cast type return
     "AddIntModule_basic",
     "SubIntModule_basic",
     "SubFloatModule_basic",
@@ -59,14 +58,7 @@ PI_XFAIL_SET = {
     "BoolIntConstantModule_basic",
     "AtenIntTensorByteDtypeModule_basic",
     "AtenIntTensorCharDtypeModule_basic",
-    "NeIntModule_basic",
-    "EqIntModule_basic",
-    "GtIntModule_basic",
-    "GeIntModule_basic",
-    "GeFloatModule_basic",
-    "GeFloatIntModule_basic",
-    "NeFloatIntModule_basic",
-    "GtFloatIntModule_basic",
+
     # type/api overload
     "ToDtypeLayoutStridedModule_basic",
     "SqueezeModule_broadcast",
@@ -89,44 +81,6 @@ PI_XFAIL_SET = {
     "ArangeFalsePinMemoryModule_basic",
     "ArangeDtypeIntModule_basic",
     "ArangeDtypeFloatModule_basic",
-    "VarMeanCorrectionNoneModule_basic",
-    "VarMeanCorrectionModule_basic",
-    "VarDimUnbiasedModule_basic",
-    "VarDimSingleDimModule_basic",
-    "VarDimNoneDimModule_basic",
-    "VarDimNegativeModule_basic",
-    "VarDimMultiDimModule_basic",
-    "VarDimModule_basic",
-    "VarDimEmptyDimModule_basic",
-    "VarDimBiasedModule_basic",
-    "VarDimAllDimReduceModule_basic",
-    "VarCorrectionSingleDimReduceModule_basic",
-    "VarCorrectionNoneModule_basic",
-    "VarCorrectionModule_basic",
-    "VarCorrectionLargeInputModule_basic",
-    "VarCorrectionKeepDimModule_basic",
-    "VarCorrectionEmptyDimModule_basic",
-    "VarCorrectionAllDimReduceModule_basic",
-    "StdDimNoneDimModule_basic",
-    "StdDimKeepDimTrueModule_basic",
-    "StdDimKeepDimFalseModule_basic",
-    "StdDimEmptyDimModule_basic",
-    "StdDimBiasedModule_basic",
-    "StdCorrectionSingleDimReduceModule_basic",
-    "StdCorrectionNoneModule_basic",
-    "StdCorrectionModule_basic",
-    "StdCorrectionLargeInputModule_basic",
-    "StdCorrectionKeepDimModule_basic",
-    "StdCorrectionEmptyDimModule_basic",
-    "StdCorrectionAllDimReduceModule_basic",
-    "MeanDimNoneDimModule_basic",
-    "MeanDimNegativeModule_basic",
-    "MeanDimModule_basic",
-    "MeanDimKeepdimModule_basic",
-    "MeanDimEmptyDimModule_basic",
-    "MeanDimDtypeModule_basic",
-    "MeanDimAllReduceModule_basic",
-    "MeanDimAllReduceKeepdimModule_basic",
     "MaxPool2dWithIndicesStaticModule_basic",
     "MaxPool2dWithIndicesNonDefaultStrideModule_basic",
     "MaxPool2dWithIndicesNonDefaultParamsModule_basic",
@@ -141,29 +95,6 @@ PI_XFAIL_SET = {
     "RandnDtypeDeviceModule_basic",
     "BernoulliModule_basic",
     "HBC_basic",
-    "ElementwiseNeIntScalarModule_basic",
-    "ElementwiseNeFloatTensorModule_basic",
-    "ElementwiseLtIntTensorModule_basic",
-    "ElementwiseLtIntScalarModule_basic",
-    "ElementwiseLtFloatTensorModule_basic",
-    "ElementwiseLtFloatScalarModule_basic",
-    "ElementwiseLtDiffWidthScalarModule_basic",
-    "ElementwiseLeMixedIntScalarModule_basic",
-    "ElementwiseLeIntScalarModule_basic",
-    "ElementwiseLeFloatScalarModule_basic",
-    "ElementwiseLeFloatIntScalarModule_basic",
-    "ElementwiseGtMixed2ScalarModule_basic",
-    "ElementwiseGtIntTensorModule_basic",
-    "ElementwiseGtIntScalarModule_basic",
-    "ElementwiseGtFloatTensorModule_basic",
-    "ElementwiseGtFloatScalarModule_basic",
-    "ElementwiseGeMixedIntScalarModule_basic",
-    "ElementwiseGeIntScalarModule_basic",
-    "ElementwiseGeFloatScalarModule_basic",
-    "ElementwiseGeFloatIntScalarModule_basic",
-    "ElementwiseEqIntTensorModule_basic",
-    "ElementwiseEqIntScalarModule_basic",
-    "ElementwiseEqDiffWidthScalarModule_basic",
     "AnyBoolTrueModule_basic",
     "AnyBoolFalseModule_basic",
     "AllBoolTrueModule_basic",
@@ -182,44 +113,6 @@ PI_XFAIL_SET = {
     "ReduceMaxAlongDim_basic",
     "ReduceMaxAlongDimNegative_basic",
     "ReduceAmaxSingleDim_basic",
-    "TypePromotionZeroRankHigherCategoryModule_basic",
-    "TypePromotionSameCategoryZeroRankWider_basic",
-    "TypePromotionSameCategoryDifferentWidthModule_basic",
-    "TypePromotionDifferentCategoryModule_basic",
-    "TypePromotionAlphaWiderModule_basic",
-    "RsubIntModule_noalpha_basic",
-    "RsubIntModule_basic",
-    "RsubFloatModule_noalpha_basic",
-    "RsubFloatModule_basic",
-    "ElementwiseWhereSelfModule_basic",
-    "ElementwiseWhereScalarSelfModule_basic",
-    "ElementwiseWhereScalarOtherModule_basic",
-    "ElementwiseWhereScalarModule_basic",
-    "ElementwiseTernaryModule_basic",
-    "ElementwiseSubScalarIntModule_basic",
-    "ElementwiseSubScalarFloatModule_basic",
-    "ElementwiseRemainderScalarModule_Int_basic",
-    "ElementwiseRemainderScalarModule_Int_Float_basic",
-    "ElementwiseRemainderScalarModule_Float_basic",
-    "ElementwiseRemainderScalarModule_Bool_basic",
-    "ElementwisePowTensorModule_basic",
-    "ElementwisePowTensorBroadcastModule_basic",
-    "ElementwisePowModule_basic",
-    "ElementwiseOrIntegerModule_basic",
-    "ElementwiseMulTensorIntModule_basic",
-    "ElementwiseMulTensorFloatModule_basic",
-    "ElementwiseMulScalarModule_int",
-    "ElementwiseMulScalarModule_float",
-    "ElementwiseMulScalarModule_basic",
-    "ElementwiseDivTensorFloatModule_basic",
-    "ElementwiseDivScalarModule_basic",
-    "ElementwiseDivRoundingModeTruncModule_basic",
-    "ElementwiseDivRoundingModeFloorModule_basic",
-    "ElementwiseAtenWhereSelfModule_basic",
-    "ElementwiseAndIntegerModule_basic",
-    "ElementwiseAddScalarIntModule_basic",
-    "ElementwiseAddScalarInt64Module_basic",
-    "ElementwiseAddScalarFloatModule_basic",
     "BroadcastToSameRankStaticModule_basic",
     "BroadcastZeroRankInputStaticModule_basic",
     "ScalarImplicitFloatModule_basic",
@@ -228,15 +121,26 @@ PI_XFAIL_SET = {
     "SortIntList_basic",
     "TransposeIntModule_basic",
     "TransposeIntNegDimsModule_basic",
+
     # error: operand types should have the same type as the list contained type
+    "IndexTensorHackedTwinModule3dInput_basic",
+    "IndexTensorHackedTwinModule_basic",
     "IndexTensorHackedTwinMultiInputNonContiguousMultipleStaticDims_basic",
+    "IndexTensorModule3dInput_basic",
+    "IndexTensorModule_basic",
+    "IndexTensorMultiInputContiguousCenter_basic",
+    "IndexTensorMultiInputContiguousOneDimDynamic_basic",
+    "IndexTensorMultiInputNonContiguousDynamic_basic",
+    "IndexTensorMultiInputNonContiguousMultipleStaticDims_basic",
+    "IndexTensorMultiInputNonContiguousOneDimDynamic_basic",
+    "IndexTensorMultiInputNonContiguous_basic",
+    "IndexTensorMultiInputOneDim_basic",
+    "IndexTensorMultiInputThreeIndexers_basic",
+    "IndexTensorMultiInput_basic",
+    "IndexTensorSelectDimModule_basic",
+
     # segfault (lol)
-    "Fill_TensorFloat64WithFloat32_basic",
-    "Fill_TensorFloat64WithFloat64_basic",
-    "Fill_TensorFloat64WithInt64_basic",
-    "Fill_TensorFloat32WithFloat32_basic",
-    "Fill_TensorFloat32WithFloat64_basic",
-    "Fill_TensorFloat32WithInt64_basic",
+
     "IndexPutImpl3DIntNonAccumulateModule_basic",
     "IndexPutImpl3DIntAccumulateModule_basic",
     "IndexPutImpl3DFloatNonAccumulateModule_basic",
@@ -257,21 +161,8 @@ PI_XFAIL_SET = {
     "CopyWithDifferentDTypesAndSizesModule_basic",
     "CopyWithDifferentDTypesModule_basic",
     "CopyWithDifferentSizesModule_basic",
+
     # eager/lazy materialization
-    "ZerosLikeModule_int",
-    "ZerosLikeModule_float",
-    "ZerosLikeModule_falsePinMemory",
-    "ZerosLikeModule_defaultDtype",
-    "OnesLikeModule_int",
-    "OnesLikeModule_float",
-    "OnesLikeModule_falsePinMemory",
-    "OnesLikeModule_defaultDtype",
-    "EmptyLikeModule_int",
-    "EmptyLikeModule_float",
-    "EmptyLikeModule_falsePinMemory",
-    "EmptyLikeModule_defaultDtype",
-    "EmptyLikeMemoryFormatModule_basic",
-    "ElementwiseAtenLogicalOrOpNegativeModule_basic",
     "LayerNormNormalizeOverAllDimsModule_basic",
     "LayerNormModule_basic",
     "LayerNormLastDimModule_basic",
@@ -279,19 +170,20 @@ PI_XFAIL_SET = {
     "BatchNorm2DModule_basic",
     "BatchNorm1DWith2DInputModule_basic",
     "BatchNorm1DModule_basic",
-    "TensorLiteralModule",
-    "AllBoolFalseModule_basic",
     "TensorIntModule_basic",
     "TensorLiteralModule_basic",
     "TensorOpaqueLiteralModule_basic",
+
     # backends
     "ConvolutionBackwardModule2DPadded_basic",
     "ConvolutionBackwardModule2D_basic",
+
     # failed to legalize operation
     "NumpyTRank0Module_basic",
     "TModuleRank0_basic",
     "TModuleRank1_basic",
     "TModuleRank2_basic",
+
     # error: found an op that was marked as backend illegal
     "AtenToDeviceModule_basic",
 }
